moves security fixes changes from oss to dev

+++ frontend/micro-ui/web/install-deps.sh
@@ -4,12 +4,17 @@ BRANCH="$(git branch --show-current)"
 
 echo "Main Branch: $BRANCH"
 
-INTERNALS="micro-ui-internals"
+if [ "$BRANCH" != "develop" ] || [ "$BRANCH" != "master" ]; then
 
-cd $INTERNALS && echo "Branch: $(git branch --show-current)" && echo "$(git log -1 --pretty=%B)" && echo "installing packages" && yarn install && echo "starting build" && yarn build && echo "building finished" && find . -name "node_modules" -type d -prune -print -exec rm -rf '{}' \; 
-cd ..
+  INTERNALS="digit-ui-internals"
+  rm -rf $INTERNALS
+  git clone -b main https://github.com/egovernments/digit-ui-internals.git $INTERNALS
+  cd $INTERNALS && echo "Branch: $(git branch --show-current)" && echo "$(git log -1 --pretty=%B)" && yarn && yarn build && find . -name "node_modules" -type d -prune -print -exec rm -rf '{}' \;
+  cd ..
 
-rm -rf node_modules
-rm -f yarn.lock
+  rm -rf node_modules
+  rm yarn.lock
+
+fi
 
 # yarn install
